[PC3/BusyIndicator] Avoid running invisible animation
Running is a property managed by a state machine. It is bound in the
transition between states. We don't run a transition to get to the first
state.

These transitions aren't needed and can be written away simplifying the
code considerably in the process.

This means krunner is invisibly animating at 60fps constantly even when
it's not doing anything.