[plasmacomponents3/TextField|TextArea] Fix placeholder text
In TextField, it was not disabled, so the mousearea that changes the
cursor shape wasn't working properly. Fixed now.

In TextArea, it was not disabled, not the right color, and was
inappropriately selectable. Fixed now.

The UI tests were testing for these conditions but were broken, and are
now fixed by these changes.

BUG: 433864
FIXED-IN: 5.80