[Wayland] Ensure pinned popup end up on the panel layer
The pinned popups need to act as a "panel extension". Easiest way to
achieve this on Wayland is to just make sure they end up on the panel
layer. This is simply done by changing the role on their corresponding
surface so that KWin treat them properly (as in like a panel but with
windows able to go under).

Neat way of doing it like this is that it will work wherever the popup
parent is coming from (a panel or the desktop).

BUG: 400317