Allow ExpandableListItem to not be expandable when it has no valid actions
Typically ExpandableListItem expands when you click on it or its
expander arrow, and shows a list of contextual actions. However there
may occasionally arise cirsumstances under which an item has no actions.
An example would be in the Disks & Devices applet when a device doesn't
match any of the Solid predicates and therefore gets no contextual
actions. In this case, we should hide the expander arrow and now let the
user try to expand it, because there would just be a little empty list
stub.