Fix Label having non-integer sizes
When using a non-integer baseLine, the height can be non-integer, resulting
in non-exact pixel alignments which generate artifacts on the screen.

This happened in Kickoff's LeaveView in Plasma 5.18 (currently used in
openSUSE Leap 15.3) where the focus outline was painted with a different
width in the top and bottom borders due to this pixel misalignment.