Turn WindowThumbnail into a texture provider
This gives a significant speed boost if windowThumbnails are then used
via a ShaderEffect without having to render to an intermediate
framebuffer.

It also gives the ShaderEffect the texture at the resolution of the
source, rather than the size of the Item.

Memory management of the node created in
WindowThumbnail::textureProvider is a bit quirky, but a pattern copied
in multiple places in Qt (QQuickFrameBufferObject for example) so should
be safe.