SvgRectsCache: do not query elements twice
I was looking into why there is so much time spent here for the logout
greeter that only has buttons. It seems like we were querying the cache
file every time even though we already have separate code that checks
whether our information is valid.

As I started looking into it, ::loadImageFromCache was %25 of the CPU
time. After this patch it's short of 1%.