Handle huge icons in Units::roundToIconSize()
This function was missing a use case for icons that are between the
sizes of huge and enormous, causing huge-sized icons to appear too large
on X11 when not using Qt scaling. This commit fixes that, and addresses
one of the sources of complaints when using scaling on X11. Small icons
are still too small, but that must and will be handled separately.

CCBUG: 356446