[calendar] Don't rely on ServiceType for distinguishing between old and new plugins
Currently we support two styles of plugins. The newer ones have their metadata in the usual KPluginMetaData format, the older ones have a different JSON format.

They are installed into the same directory and we assume that the new ones have "ServiceType: Plasma/CalendarPlugin".

However the servicetype is technically not needed to load the plugin so we started to remove them, e.g. in https://invent.kde.org/pim/kdepim-addons/-/commit/95186ed513a76402b8539f833ed7beed4fdcb799

This results in new-style plugins being loaded as old-style plugins, which breaks displaying the name in the config dialog.

Instead check if the metadata contains a "KPlugin" entry to determine whether a plugin is new-style or old-style