avoid holding onto old Svg object when changing source of an IconItem
A long-lived IconItem instance can have its source changed many times
over its lifetime. Because SvgSource parents its internal Plasma::Svg
instance to the IconItem instance, this means that such Plasma::Svg
instance was not being destroyed when its responsible SvgSource
instance is destroyed and indeed would not be destroyed until the
IconItem instance is eventually destroyed, which could be arbitrarily
much later. This commit adds an explicit call in the SvgSource
destructor to delete the Plasma::Svg instance so it does not hang
around in memory until the IconItem instance is destroyed. This fixes
one of the major memory "leaks" in plasmashell.