Revert "Change busywidget to a gear"
This reverts commit aafe6523cc0fe809d03fb6b98c624ac3c0f49973.

This change has not been well-received by users, and exposes a
conceptual issue with how app launch animations are currently handled in
Plasma, with the loading spinner simply re-used as an overlay on top of
the app icon. This constrains possible designs for the spinner because
the backgrounds it will be displayed on top of differ gratly in these
usages; what looks good in one context may not look good in the other.
We will have to find another way.

BUG: 442525
FIXED-IN: 5.87