ScrollView: Do not overlay scrollbars over contents
This ensures the scrollbar of a ScrollView do not overlap the ScrollView's
contents. It effectively makes ScrollView's behaviour the same as all
widget views, which also do not overlap the contents.

It also removes the need for a lot of workarounds in applications and other
places to fix the overlapping problem.

CCBUG: 443173