Corona: save after ending edit mode
Plasma can be somewhat amnesiac if you make changes to it that don't
involve adding/removing panels and it subsequently unexpectedly ends,
such as when it or the compositor crashes.

Calling requireConfigSync after edit mode ends guarantees that any changes
made by the user in edit mode will be applied.