PC3/Breeze sliders: add solid background, handle shadow, fix pixel alignment, fix vertical groove height
Also reduce opacity when disabled and use layer.enabled when opacity < 1
so that the background doesn't show through the handle.

Given what I know about QPalette, setting the opacity for the whole
control when disabled seems like it should be unnecessary. For now, this
is fine since sliders don't normally contain anything besides the groove
and handle graphics.