Introduce dismissOnHoverLeave property and toolTipVisibleChanged signal
Introduces new signal `toolTipVisibleChanged(bool toolTipVisible)`, so users can be
signalled when the tooltip is shown or hidden.

This is needed for implementing a fix to https://bugs.kde.org/show_bug.cgi?id=444142,
so that tooltips opened with a click are not dismissable by hovering over another Task.

CCBUG: 444142