BackgroundMetrics: Remove extra Window.window (is not null) condition
Casting something to boolean as a guard before the instanceof check is
redundant and can be safely removed, because even is a window is null
(e.g. not yet initialized or an item just isn't in a QQuickWindow's
content item tree hierarchy) the expression `(null) instanceof Something`
is always false anyway. No runtime errors are emitted from such checks.

GIT_SILENT