PlasmoidHeading: Turn location into an alias for position property
It's deprecated, but it can't be removed for compatibility reasons.
It's enum valued, but the values are identical to those of T.ToolBar,
so they are perfectly compatible, and it shouldn't be a breaking change
to replace one with another.

Default value of position property is Header, which is set in C++ in
constructor. Specifying it in QML have used to have the negative
side-effect of breaking the magic of automatic (implicit) positioning
done by TabBar, ToolBar and DialogButtonBox.