[WIP] Deprecate PlasmaCore.Units.largeSpacing
It has to be stopped, and every current usage ported to gridUnit or
2*smallSpacing, whatever suits the context. Because it is inconsistent
with Kirigami.Units, and just equals to gridUnit all the time, so this
property does not really make any sense, and is only ever confusing.

Note: this patch does not compile, due to the fact that units.{h,cpp}
files are shared between two CMakeLists.txt in different
adjacent directories. Probably worth splitting them into their own target?