Don't crash when a screen gets disabled and enabled again
when a screen gets disabled, the containment loses its view, and when gets enabled again,
the process of the old orphaned containment getting a new view causes a plasma crash.
It comes from the wallpaper, which indicates a hierarchy of C++ owned objects
makes the internal one (wallpaper) to not reset correctly its window and have a stray pointer to it.

BUG:447752