Port away from QXcbWindowFunctions for Qt6
The challenge here is that QXcbWindow::setWmWindowType is no longer
accessible to us, due to the corresponding header not being installed.
Looking at the implementation however shows it's reading the WM type also
from a the "_q_xcb_wm_window_type" property on the window, something
KCompletion is apparently already using.

Obviously I have no way to test this yet, it is possible that we do not
trigger the code reading the property in all cases yet.