Hardcode keyboard shortcut in SearchField for now
StandardKey.Find does not work in this context, and instead becomes
translated into the text "22", which its its enum's int value as per
https://doc.qt.io/qt-5/qkeysequence.html#StandardKey-enum.

As a result, hitting Ctrl+F does not focus the field, and hovering over
it with the cursor shows a tooltip with only "22" in it.

Until this can be investigated and fixed, let's hardcode "Ctrl+F"  for
now to unbreak it, which is what the Kirigami version does too.

CCMAIL: carl@carlschwan.eu