Fix Plasmoid attached property look-up on Config objects
At the moment we were only keepy track of them for the applets
themselves and propagating the context property.

Instead, now we are storing the value in the root context and pulling it
from there for every object. We are doing it too for Config files right
where we are also forwarding the context property.

This changes how the context property is stored by using dynamic
properties in the context object instead of a global hash. This has the
advantage of doing without a global vector that is getting accessed from
different places.

BUG: 451482