Highlight: change hover: property to hovered:
The PlasmaComponents2 version of this item had a property named
hovered:. When a PlasmaExtras version was made, the property name was
changed to hover:, which is arguably a worse name since it no longer
communicates that it's a state that you might need to set manually.

Changing the name also prevents 1:1 ports since you'll have to know
the difference and can easily forget it. This happened once already in
https://invent.kde.org/frameworks/plasma-framework/-/merge_requests/486

The component hasn't been released yet, so it's not too late to change
the property name. Let's do it now.