ExpandableListItem: deprecate contextmenu and allow actions+custom view
The purpose of this component is to provide visible,
left-click/touch-friendly access to actions contextually relevant to an
item. Allowing items to define context menus was a mistake, as the
existing paradigm of expanding the item to show actions provides the
same functionality in a more consistent and user-friendly way that stays
truer to the component's design.

This commit deprecates the contextMenu: property and allows contextual
items to be defined as well as custom context views; when both are
defined, actions are simply shown above the custom view.

There are no breaking API or ABI changes, and no visual changes to
existing applets that don't define both a contextual actions lists
and a custom context menu.