Automatically add tooltips and Accesible properties for icon-only things
Currently we make icons-only buttons in a mishmash of ways, and
sometimes forget to manually add tooltips and accessibility properties
to them. A better approach is to do this automatically using the text:
property for the button, and use the button's display: property to
denote that it should be an icons-only button. This will let us get rid
of loads of custom code in our UI controls and ensure that tooltips and
accessibility metadata are always present.