Draft: Set currentIndex even when hovering over nested buttons
This one almost works, except HoverHandler is buggy, and won't start
tracking a point if it immediately jumps straight on top of a nested
control thus bypassing direct interaction with HoverHandler. However,
once a handler is hovered directly, it will react to a pointer even on
top of nested controls.