PC3/BusyIndicator: Center the active/rotating part of the control
Otherwise, with non 1:1 aspect ratios it would stick to top or left. For
example, this problem can be observed by adding an icon applet to a
panel, and launching it: busy indicator of the applet would not be
canter-aligned to the icon, because of added padding making the applet
container a slightly squashed rectangle.