Never time out tooltip while its control is still hovered
Right now there's logic to try to predict how long a user will need to
read the tooltip, and time it out after that. This is unreliable and
requires making an assumption about the user's reading speed, which is
unknowable. As a result, the tooltip can still time out while you're in
the middle of reading it, which is annoying.

There's no real reason why we need to time it out automatically at all,
so let's just keep it open while the user is still hovering the control.
They're probably doing so for a good reason.

BUG: 397336
FIXED-IN: 5.100