ExpandableListItem: Fix warning
It would appear on the wrong object, as we are trying to call the wrong
object's attached property.
org.kde.plasma.bluetooth/contents/ui/DeviceItem.qml:33:32: QML Action: Accessible must be attached to an Item

Changed the Action's text because otherwise it goes unused. Using name
instead of description because it fits better the documentation.