declarative/core: Use two opacity nodes for fading in IconItem
There's no need to create a complete custom node with material and
shaders if all we're doing is a linear fade between the old image and
the new. Instead we can use two QSGOpacityNodes for the fading, which
gives us the same effect without the need for so much custom code.

Moreover, at least one of the duplicates of 432999 points to problems in
fadingnode, so this hopefully reduces the changes of that occuring.

CCBUG: 432999