Rework Plasma theme mask corners and outlines
This patch reverts Marco's idea of making the mask of dialogs smaller by 1px
(and https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/2276 reverts
my patch doing the same for the panel); instead, this MR reduces the mask by 1px,
but only for the corner elements.

The issue with reducing the whole mask by 1px is that you'll get a dark outline in
Breeze Dark (both on the panel and the dialogs) *and* you'll be able to see the
wallpaper through the panel/dialog in its outermost pixel.

This patch still isn't perfect. With some combination of colorscheme and wallpaper,
you might still notice some slight darkening going on in the corners. This happened
before this MR too, but at least now it happens *only* in the corners. It's much
less noticeable; especially since most people seem to notice it only for the panel,
which in most cases does not have any rounded corners (but it does have sides, so
this MR - which makes the bug only appear in corners - is a great step forward).

BUG: 417511
FIXED-IN: 5.101