Extra/SearchField: Clear the search field on Escape pressed
Otherwise the event is treated at a higher level, usually
to dismiss a dialog or a full representation, which may be
unexpected or annoying for the user.

Just do that if the user already started typing something.
If the search field is empty we pass the event forward
(as previously) so a quick dismiss is still available.