Mark transient applets in config
When deleting an applet it doesn't really delete it but sets it invisible and marks it
as "transient" to be able to undo the deletion.

If plasma is killed or crashes, the information that it was deleted is lost and Plasma
recreates it on the next launch. This commit saves the "transient" property in the
config and on launch actually cleans the config and doesn't load those.

Not actually deleting the config groups on first delete, as it would cause config loss
on undo.

BUG: 417114
FIXED-IN: 5.102