iconitem: Adjust fade animation to not flicker as much
As it turns out, mix(old texture, new texture, animation value) does
actually result in a different visual from simply blending the two
textures on top of each other, because the old texture first gets
blended with the underlying color. This results in some of the
underlying color bleeding through, which can cause flickering if the two
textures are almost the same. To avoid this, change the animation
slightly so that rather than crossfading, we first fade in the new
texture then fade out the old texture. This doesn't cause a whole lot of
visual difference for two completely different textures, but removes the
flickering when they are very similar.

BUG: 463061
BUG: 463685