IconItemTest: Wait for rendering before capturing an image of IconItem
When animated, we want to wait a bit before trying to capture an image
of IconItem as otherwise the animation has had no time to run and
nothing will have changed yet. Additionally, we want to make sure the
window we're rendering to is actually visible as we otherwise risk
nothing being rendered.