Dialog: Fix calculation of distance between dialog and parent applet
`ParentRect` is not using the same coordinate system as `dialogPos` and `avail`,
which could result in incorrect placement of dialogs when you have e.g. multiple
monitors or a non-maximized panel. This replaces `ParentRect` with `dialogPos` so
the coordinate systems remain in sync.

BUG:464201
BUG:464513
FIXED-IN: 5.103


(cherry picked from commit eda6caabcffff484544af8949d0ca0a3ea3ce333)