PC3 menu: make interactive condition more robust
If the menu has a content height that exceeds the available height, but
ApplicationWindow.window isn't set--for example because we're in the
SDDM login screen where there is no application window--then we'll hit
the fallback condition and be non-interactive, which makes the menu
terribly difficult to scroll with a pointing device and impossible to
scroll with a touchscreen.

To fix this, consult Window.window rather than ApplicationWindow.window.
This is what other styles do.

BUG: 466968
FIXED-IN: 5.105


(cherry picked from commit e4830f160bcec9e30f6181c75dead7817020d874)