Bring back contextual actions
Contextual actions depended on appletscript.
what the old thing did was taking all the actions accessed by
AppletInterface::action("name") by qml (but not c++) and adding them
in the context menu.
This restores 100% of the behavior with an ad hoc property set by
appletinterface

This will have to be overhauled when Plasmoid will have to
become Applet *