Fix osd- icon IDs
Multi-size icons need for the *largest* size to omit a size prefix, and
then all of the smaller versions have a size prefix as a hint to the
system. This was wrong for the osd- icons, which used the 64px version
as the un-prefixed size. To fix that, un-prefix the largest size as
expected.

BUG: 469303
FIXED-IN: 5.106
(cherry picked from commit 4f39497fc901f01c07b3a4cc37aaf8c44090dea1)