Reintroduce immutable property
For compatibility, reintroduce a Q_PROPERTY called immutable
which is true whether the immutability is not mutable.

This makes it compatible with the existing QML, for which it's
a clearer API to be just a bool. If needed for more granularity
between UserImmutable and SystemImmutable, a new property can
be added later, but it's still better to have a simpler bool
by default.