Use KSvg::ImageSet to load themes
This patch adds a KSvg dependency and instantiates KSvg::ImageSet and sets it to load the current plasma theme, used internally in Plasma::Theme, which is the only svg-related thing that is going to remain in plasma-framework.

This is a non breaking change as plasmoids can start to be ported and svg/framesvg can be removed from here only after porting is done.