Build plumbing for allowing applet config pages to omit internal margins
Right now all applet config pages automatically get margins and this
can't be disabled. Most of the time this is fine. But sometimes, you do
want to disable them because the page itself wants to have a frameless
style, with an inner list view that touches the edges of the window and
sidebar. Currently this can't be done.

With this commit, the infrastructure now supports pages registering
that they want to opt out of automatic margins by setting
`includeMargins: false` in their config.qml files where they define
their config pages. They will then be responsible for adding any needed
margins themselves.