SwitchIndicator: Enable layered rendering when semi-transparent
This is the trick we use for Slider as well. Without a dedicated layer,
parts of this component are rendered semi-transparent individually, so
that the ones below are "shining through" the ones above them. For
example, in this case you could see disabled switch indicator's groove
under the knob.

(cherry picked from commit ebe9bd91e5edf02d867dd4916f5ddd969043bc1d)