Dialog: Only perform setup in the first expose event after showing a window
There is a code path to update the intiial visibility on wayland when we
hit an expose event. It's only processed on wayland so some of this code
was dead code.

It's important this only runs once as it updates the theme and
visibility, which for badly written contents can infinitely loop.

To restore behaviour as close as possible to the current state we use a
boolean flag to make sure we only process expose once.