Initalise PlasmaShellManager in it's ctor
QtWaylandClinetExtensionTemplate can load everything on startup, it
shares the registry with QGuiApplication.

However because it calls a virtual function in initialize it can't be
called from the base class constructor. It needs to be called in the
leaf class, or explicitly after construction otherwise it falls back to
the default implementation which is a metaobject invoke. A Dialog could
be created before that triggers.

BUG: 474452