BasicPlasmoidHeading: Subscribe to "configure" action updates
Some applets replace "configure" action with their custom thing. For
such applets is it important either to have strong guarantees that the
action will be fetched by a ToolButton strictly after the replacement,
or have the button subscribe to changes at runtime (which is way more
reliable in any case).