Support PlasmoidHeading on other than Dialog and AppletPopup
The new panel config ui uses a PlasmoidHeading in a PopupPlasmaWindow
which is not exposed in QML, so the logic for the margins fails and
doesn't render correctly.

This makes it use the padding properties of the window when available.
Unfortunately uses duck-typing as the relevant type is not exposed,
in order to have something more type safe PopupPlasmaWindow should be
binded to QML, (even if uncreatable) tough it would probably be more
weird than this

before:

![Screenshot_20231013_165302](/uploads/ad684104c8a02dae5ab6218ab9140a5a/Screenshot_20231013_165302.png)

after:

![Screenshot_20231013_164525](/uploads/4e45352bc62a33188d4be0218fa98011/Screenshot_20231013_164525.png)