Highlight: Add API to let callers use an "inactive highlight" effect
A pattern used in a variety of widgets is to have an "inactive
highlight" effect. This is for items that are marked as hovered or
selected when something else with focus is currently hovered or
selected; the inactive thing keeps the same visual state, but with
reduced opacity.

All widgets that currently use this paradigm roll their own
implementation of it. Let's make it possible to do it centrally by
adding some new API.

Part of https://invent.kde.org/plasma/plasma-workspace/-/issues/97.