waylandintegration: Avoid creating shell surfaces on teardown
The original code intended to call platformSurfaceChanged on both
surface creation and destruction, but use the presence of a native
handle as a way to only track creation.

This failed as the opposite event to window creation is
aboutToBeDestroyed, and the native window handle is still valid.

Instead of checking the presence of a native window, use the type from
the event.