ExpandableListItem: Narrow actions type down to T.Action
Templates are the correct types when declaring properties, because they
are not bound to any particular style implementation/override.

All known users of this property either declare their action as
QQC2.Action or Kirigami.Action, or in case of Disks & Devices applet it
converts QAction list to QQC2.Action via Instantiator anyway.